pyproject.toml
[project]
name = "expensetrackerapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
"aioboto3>=14.3.0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "authlib>=1.5.2",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.1.0",
    "resend>=2.10.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]
